Do not create siblings when updating bucket ACL.
Fixes #166

Stanchion was creating a brand new object instead of updating the
existing object. As a result the put did not include the vclock for
the existing object and a sibiling resulted if allow_mult was true on
the moss.buckets bucket. This fix changes the behavior to update the
existing object for ACL updates and avoids the sibling creation.